Recognize args inside macro definition with the help of the scanner
Unfortunately, current syntax is very inconsistent. So, while normally
you can have arguments with a comma (,), when inside a macro definition
you can't, because comma is used as a separator between multiple
commands. For this reason we need a scanner that is able to do things
differently based on the fact we are inside a macro definition or not.